LPD-51848 Rename file
